+++ R/jsBasicGadget.R
@@ -6,7 +6,7 @@
 #' @details Shiny Gadget including Data, Label info, Table 1, Regression(linear, logistic), Basic plot
 #' @examples
 #' if (interactive()) {
-#'   jsBasicGadjet(mtcars)
+#'   jsBasicGadget(mtcars)
 #' }
 #' @rdname jsBasicGadget
 #' @export
@@ -267,6 +267,88 @@ jsBasicGadget <- function(data, nfactor.limit = 20) {
             withLoader(tableOutput("cut_timeroc"), type = "html", loader = "loader6"),
             ggplotdownUI("timeroc"),
             withLoader(DTOutput("table_timeroc"), type = "html", loader = "loader6")
+          )
+        )
+      )
+    ),
+    navbarMenu(
+      title = "Subgroup analysis",
+      icon = icon("chart-bar"),
+      tabPanel(
+        title = "subgroup cox",
+        sidebarLayout(
+          sidebarPanel(
+            forestcoxUI("Forest")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot", width = "100%"),
+                ggplotdownUI("Forest")
+              )
+            )
+          )
+        )
+      ),
+      tabPanel(
+        title = "subgroup regression",
+        sidebarLayout(
+          sidebarPanel(
+            forestglmUI("Forest_glm")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub_glm"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot_glm", width = "100%"),
+                ggplotdownUI("Forest_glm")
+              )
+            )
+          )
+        )
+      ),
+      tabPanel(
+        title = "subgroup logistic regression",
+        sidebarLayout(
+          sidebarPanel(
+            forestglmUI("Forest_glmbi")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub_glmbi"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot_glmbi", width = "100%"),
+                ggplotdownUI("Forest_glmbi")
+              )
+            )
           )
         )
       )
@@ -655,6 +737,28 @@ jsBasicGadget <- function(data, nfactor.limit = 20) {
       caption.placement = "top"
     )
 
+
+    outtable <- forestcoxServer("Forest", data = data, data_label = data.label)
+    output$tablesub <- renderDT({
+      outtable()[[1]]
+    })
+    output$forestplot <- renderPlot({
+      outtable()[[2]]
+    })
+    outtable_glm <- forestglmServer("Forest_glm", data = data, data_label = data.label, family = "gaussian")
+    output$tablesub_glm <- renderDT({
+      outtable_glm()[[1]]
+    })
+    output$forestplot_glm <- renderPlot({
+      outtable_glm()[[2]]
+    })
+    outtable_glmbi <- forestglmServer("Forest_glmbi", data = data, data_label = data.label, family = "binomial")
+    output$tablesub_glmbi <- renderDT({
+      outtable_glmbi()[[1]]
+    })
+    output$forestplot_glmbi <- renderPlot({
+      outtable_glmbi()[[2]]
+    })
     session$onSessionEnded(function() {
       stopApp()
     })
@@ -1000,6 +1104,88 @@ jsBasicExtAddin <- function(nfactor.limit = 20, max.filesize = 2048) {
           )
         )
       ),
+    ),
+    navbarMenu(
+      title = "Subgroup analysis",
+      icon = icon("chart-bar"),
+      tabPanel(
+        title = "subgroup cox",
+        sidebarLayout(
+          sidebarPanel(
+            forestcoxUI("Forest")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot", width = "100%"),
+                ggplotdownUI("Forest")
+              )
+            )
+          )
+        )
+      ),
+      tabPanel(
+        title = "subgroup regression",
+        sidebarLayout(
+          sidebarPanel(
+            forestglmUI("Forest_glm")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub_glm"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot_glm", width = "100%"),
+                ggplotdownUI("Forest_glm")
+              )
+            )
+          )
+        )
+      ),
+      tabPanel(
+        title = "subgroup logistic regression",
+        sidebarLayout(
+          sidebarPanel(
+            forestglmUI("Forest_glmbi")
+          ),
+          mainPanel(
+            tabsetPanel(
+              type = "pills",
+              tabPanel(
+                title = "Data",
+                withLoader(
+                  DTOutput("tablesub_glmbi"),
+                  type = "html",
+                  loader = "loader6"
+                )
+              ),
+              tabPanel(
+                title = "figure",
+                plotOutput("forestplot_glmbi", width = "100%"),
+                ggplotdownUI("Forest_glmbi")
+              )
+            )
+          )
+        )
+      )
     )
   )
 
@@ -1235,7 +1421,28 @@ jsBasicExtAddin <- function(nfactor.limit = 20, max.filesize = 2048) {
       caption = "Best cutoff",
       caption.placement = "top"
     )
+    outtable <- forestcoxServer("Forest", data = data, data_label = data.label)
 
+    output$tablesub <- renderDT({
+      outtable()[[1]]
+    })
+    output$forestplot <- renderPlot({
+      outtable()[[2]]
+    })
+    outtable_glm <- forestglmServer("Forest_glm", data = data, data_label = data.label, family = "gaussian")
+    output$tablesub_glm <- renderDT({
+      outtable_glm()[[1]]
+    })
+    output$forestplot_glm <- renderPlot({
+      outtable_glm()[[2]]
+    })
+    outtable_glmbi <- forestglmServer("Forest_glmbi", data = data, data_label = data.label, family = "binomial")
+    output$tablesub_glmbi <- renderDT({
+      outtable_glmbi()[[1]]
+    })
+    output$forestplot_glmbi <- renderPlot({
+      outtable_glmbi()[[2]]
+    })
     session$onSessionEnded(function() {
       stopApp()
     })