+++ NEWS.md
@@ -1,3 +1,10 @@
+# jsmodule 1.5.1
+
+## Update: Add Subgroupanalysis to jsBasicGadget
+## Update: Add figure in subgroup analysis with forestplot
+## Update: Add family 'poisson' in forestglm
+## Update: Add "line size" "point size" to lineplot
+
 # jsmodule 1.5.0
 
 ## New module: Subgroup analysis with forestplot (Thanks for [Yoonkyoung Jeon](https://github.com/cyk0315))